project folder/requirements.txt
accelerate>=1.12.0
agno>=2.3.20
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.2
aiosignal>=1.4.0
aiosqlite>=0.22.0
annotated-doc>=0.0.4
annotated-types>=0.7.0
anyio>=4.12.0
attrs>=25.4.0
backoff>=2.2.1
bcrypt>=5.0.0
beautifulsoup4>=4.14.3
build>=1.3.0
cachetools>=6.2.4
cbor>=1.0.0
certifi>=2025.11.12
cffi>=2.0.0
charset-normalizer>=3.4.4
chromadb>=1.3.7
click>=8.3.1
colorama>=0.4.6
coloredlogs>=15.0.1
cryptography>=46.0.3
dataclasses-json>=0.6.7
datasets>=4.4.2
dill>=0.4.0
distro>=1.9.0
dnspython>=2.8.0
docstring_parser>=0.17.0
durationpy>=0.10
ecdsa>=0.19.1
email-validator>=2.3.0
eval_type_backport>=0.3.1
fastapi>=0.115.5
filelock>=3.20.1
FlagEmbedding>=1.3.5
flatbuffers>=25.12.19
frozenlist>=1.8.0
fsspec>=2025.10.0
gitdb>=4.0.12
GitPython>=3.1.45
google-ai-generativelanguage>=0.6.15
google-api-core>=2.28.1
google-api-python-client>=2.187.0
google-auth>=2.45.0
google-auth-httplib2>=0.3.0
google-generativeai>=0.8.5
googleapis-common-protos>=1.72.0
greenlet>=3.3.0
grpcio>=1.76.0
grpcio-status>=1.71.2
h11>=0.16.0
h2>=4.3.0
hpack>=4.1.0
httpcore>=1.0.9
httplib2>=0.31.0
httptools>=0.7.1
httpx>=0.28.1
httpx-sse>=0.4.3
huggingface-hub>=0.36.0
humanfriendly>=10.0
hyperframe>=6.1.0
idna>=3.11
ijson>=3.4.0.post0
importlib_metadata>=8.7.1
importlib_resources>=6.5.2
inscriptis>=2.7.0
invoke>=2.2.1
ir_datasets>=0.5.11
Jinja2>=3.1.6
joblib>=1.5.3
jsonpatch>=1.33
jsonpointer>=3.0.0
jsonschema>=4.25.1
jsonschema-specifications>=2025.9.1
kubernetes>=34.1.0
langchain-classic>=1.0.0
langchain-community>=0.4.1
langchain-core>=1.2.4
langchain-text-splitters>=1.1.0
langsmith>=0.5.0
lxml>=5.4.0
lz4>=4.4.5
markdown-it-py>=4.0.0
MarkupSafe>=3.0.3
marshmallow>=3.26.2
mdurl>=0.1.2
mistralai>=1.10.0
mmh3>=5.2.0
mpmath>=1.3.0
multidict>=6.7.0
multiprocess>=0.70.18
mypy_extensions>=1.1.0
networkx>=3.6.1
numpy>=2.4.0
oauthlib>=3.3.1
onnxruntime>=1.23.2
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
# opentelemetry-semantic-conventions>=0.60b1
orjson>=3.11.5
overrides>=7.7.0
packaging>=25.0
pandas>=2.2.2
passlib>=1.7.4
peft>=0.18.0
posthog>=5.4.0
propcache>=0.4.1
proto-plus>=1.27.0
protobuf>=5.29.5
psutil>=7.1.3
pyarrow>=22.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pybase64>=1.4.3
pycparser>=2.23
pydantic>=2.10.3
pydantic-settings>=2.12.0
pydantic_core>=2.27.1
Pygments>=2.19.2
pyparsing>=3.3.0
PyPika>=0.48.9
pyproject_hooks>=1.2.0
pyreadline3>=3.5.4
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-jose>=3.5.0
python-multipart>=0.0.21
pytz>=2025.2
PyYAML>=6.0.3
referencing>=0.37.0
regex>=2025.11.3
requests>=2.32.5
requests-oauthlib>=2.0.0
requests-toolbelt>=1.0.0
rich>=14.2.0
rpds-py>=0.30.0
rsa>=4.9.1
safetensors>=0.7.0
scikit-learn>=1.8.0
scipy>=1.16.3
sentence-transformers>=5.2.0
sentencepiece>=0.2.1
setuptools>=80.9.0
shellingham>=1.5.4
six>=1.17.0
smmap>=5.0.2
sniffio>=1.3.1
soupsieve>=2.8.1
SQLAlchemy>=2.0.45
starlette>=0.41.3
sympy>=1.14.0
tenacity>=9.1.2
threadpoolctl>=3.6.0
tokenizers>=0.22.1
torch>=2.9.1
tqdm>=4.67.1
transformers>=4.57.3
trec-car-tools>=2.6
typer>=0.20.1
typing-inspect>=0.9.0
typing-inspection>=0.4.2
typing_extensions>=4.12.2
tzdata>=2025.3
unlzw3>=0.2.3
uritemplate>=4.2.0
urllib3>=2.3.0
uuid_utils>=0.12.0
uvicorn>=0.32.1
warc3-wet>=0.2.5
warc3-wet-clueweb09>=0.2.5
watchfiles>=1.1.1
websocket-client>=1.9.0
websockets>=15.0.1
xxhash>=3.6.0
yarl>=1.22.0
zipp>=3.23.0
zlib-state>=0.1.10
zstandard>=0.25.0
